COL-1486, create-course at LTI launch fix; move merged/lti code to more sensible spots
