+++ squiggy/api/canvas_controller.py
@@ -24,21 +24,65 @@ ENHANCEMENTS, OR MODIFICATIONS.
 """
 
 from flask import current_app as app, request, Response
-from squiggy.merged.lti import get_lti_cartridge_xml
+from squiggy.lib.lti import get_tool_metadata, TOOL_ID_ASSET_LIBRARY, TOOL_ID_ENGAGEMENT_INDEX
 
 
 @app.route('/lti/cartridge/asset_library.xml')
 def asset_library_xml():
-    return _get_lti_cartridge(tool_id='suitec:asset_library')
+    return Response(
+        _get_lti_cartridge_xml(
+            host=request.headers['Host'],
+            tool_id=TOOL_ID_ASSET_LIBRARY,
+        ),
+        mimetype='application/xml',
+    )
 
 
 @app.route('/lti/cartridge/engagement_index.xml')
 def engagement_index_xml():
-    return _get_lti_cartridge(tool_id='suitec:engagement_index')
-
-
-def _get_lti_cartridge(tool_id):
     return Response(
-        get_lti_cartridge_xml(host=request.headers['Host'], tool_id=tool_id),
+        _get_lti_cartridge_xml(
+            host=request.headers['Host'],
+            tool_id=TOOL_ID_ENGAGEMENT_INDEX,
+        ),
         mimetype='application/xml',
     )
+
+
+def _get_lti_cartridge_xml(host, tool_id):
+    tool_metadata = get_tool_metadata(host=host, tool_id=tool_id)
+    launch_url = tool_metadata['launch_url']
+    title = tool_metadata['title']
+    return f"""<?xml version="1.0" encoding="UTF-8"?>
+        <cartridge_basiclti_link
+          xmlns="http://www.imsglobal.org/xsd/imslticc_v1p0"
+          xmlns:blti = "http://www.imsglobal.org/xsd/imsbasiclti_v1p0"
+          xmlns:lticm ="http://www.imsglobal.org/xsd/imslticm_v1p0"
+          xmlns:lticp ="http://www.imsglobal.org/xsd/imslticp_v1p0"
+          xmlns:xsi = "http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation = "
+            http://www.imsglobal.org/xsd/imslticc_v1p0 http://www.imsglobal.org/xsd/lti/ltiv1p0/imslticc_v1p0.xsd
+            http://www.imsglobal.org/xsd/imsbasiclti_v1p0 http://www.imsglobal.org/xsd/lti/ltiv1p0/imsbasiclti_v1p0.xsd
+            http://www.imsglobal.org/xsd/imslticm_v1p0 http://www.imsglobal.org/xsd/lti/ltiv1p0/imslticm_v1p0.xsd
+            http://www.imsglobal.org/xsd/imslticp_v1p0 http://www.imsglobal.org/xsd/lti/ltiv1p0/imslticp_v1p0.xsd">
+          <blti:title>{title}</blti:title>
+          <blti:description>{tool_metadata['description']}</blti:description>
+          <blti:launch_url>{launch_url}</blti:launch_url>
+          <blti:extensions platform="canvas.instructure.com">
+            <lticm:property name="tool_id">{tool_id}</lticm:property>
+            <lticm:property name="privacy_level">public</lticm:property>
+            <lticm:options name="course_navigation">
+              <lticm:property name="url">{launch_url}</lticm:property>
+              <lticm:property name="text">{title}</lticm:property>
+              <lticm:property name="visibility">public</lticm:property>
+              <lticm:property name="default">disabled</lticm:property>
+              <lticm:property name="enabled">false</lticm:property>
+              <lticm:options name="custom_fields">
+                <lticm:property name="external_tool_url">$Canvas.externalTool.url</lticm:property>
+              </lticm:options>
+            </lticm:options>
+          </blti:extensions>
+          <cartridge_bundle identifierref="BLTI001_Bundle"/>
+          <cartridge_icon identifierref="BLTI001_Icon"/>
+        </cartridge_basiclti_link>
+    """
